requirements.txt
openai>=1.30.0,<2.0
pyyaml>=6.0
jinja2>=3.1